Split SourceLink provenance implementation
